x86/resctrl: Move ctrlval string parsing links away from the arch code
How to parse the ctrlval strings from user-space is specified by a
function pointer the arch code specifies.

The strings are part of resctrl's ABI, and the functions and their caller
both live in the same file. Exporting the parsing functions and allowing
the architecture to choose allows another architecture to get this wrong.

Keep this all in the flesystem parts of resctrl. This should prevent any
architecture's string-parsing behaviour from varying without core code
changes. Use the fflags to spot caches and bandwidth resources, and use
the appropriate helper.

Signed-off-by: James Morse <james.morse@arm.com>